+++ CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ### Added
 - Added new `setDefaultMemberRole` mutation to member roles resolver.
+- Added `defaultTemplate` query to the `versionedTemplate` schema and a corresponding resolver
+- Added `markAsDefaultTemplate` mutation to the `template` resolver (superadmin only)
+- Added `setDefaultTemplate` function to the `TemplateService`
+- Added `isDefault` flag to `Template` and `VersionedTemplate`
+- Added a `addAlternateIdentifierToPlan` and `removeAlternateIdentifierFromPlan` resolvers and schema to facilitate adding and removing alternate identifiers for a Plan
+- Added a `planByAlternateIdentifier` resolver and schema to facilitate finding Plans by a alternate id
+- Added `AlternateIdentifier` table and model to store alternate identifiers for Plans/DMPs
+- Added new `resolveNamingCollision` helper method and updated `Plan.create` to use it to prevent duplicate plan names. 
 - Added local seed data for a Template that includes a research output question type, and a project/plan that uses it.
 - Added `PRIMARY` ProjectCollaboratorAccessLevel [#227]
 - Added `findByUserIdAndProjectId function to Collaborator model [#227]
@@ -72,6 +80,8 @@
 
 ### Updated
 - Updated `MemberRole` schema to include `isDefault` flag and added a `setDefaultMemberRole` mutation.
+- Updated `TemplateService` to ensure that the `isDefault` flag is propagated when creating a template version
+- Updated `Plan` schema to include `alternateIdentifiers` as a chained resolver.
 - Refactored the `saveMaDMPVersion` function in `planService` to use the shared functionality from `@dmptool/utils` package to write directly to Dynamo rather than sending SQS messages.
 - Updated `updateProjectCollaborator` resolver to use new collaborator service functions to validate access level change and to demote an existing `primary` permission, before promoting another collaborator [#227]
 - Updated permissions for who can `requestFeedback` [#227]
@@ -167,6 +177,7 @@
 - Removed `ioredis` package
 
 ### Fixed
+- Fixed an issue with duplicate URIs being returned to `findRe3DataByURIs` by deduping [#33]
 - Fixed bug in `openSearchService` that was throwing an error and not returning repositories [#196]
 - Fixed some new `type` errors in `feedback` resolver and `emailService` brought on by a recent update to `typescript-eslint` [#189]
 - Had issue running `nuke-db.sh` and `process.sh`, so I turned off SSL by using `--ssl=off` instead